--- src/main/menu/subMenuFile.js
@@ -1,8 +1,0 @@
-const subMenuFile = {
-  label: 'File',
-  submenu: [
-    { role: 'quit' }
-  ]
-};
-
-export default subMenuFile;
